frameworks/native: Do Binder IPC for transform hint only when necessary.
Transform hint can give opengl-es client hint to
rotate surface so that extra composition step to rotate
surface is avoided.
IPC call is made everytime for getting transform-hint.
This IPC is unnecessary if hint is not going to change
in app's life-cycle. IPC is needed only when
transform-hint (updated in Q) is obselete for a frame,
which can happen in DQ DQ and Q sequence.
Call IPC only for DQ DQ and Q sequence.

Change-Id: I09e8326f1483434248423c1e6d7b1848aa0b95a2