sf: Refactoring of Virtual display to HWC
Refactoring is done to address framework reboot issue when
virtual display is connected and disconnected back to back.

Cleaning up of displaydevice resources as part of eDisplayTransaction
is asynchronous and sometimes there is a significant delay between
the point at which it is invoked and to the point it is being freed.
Due to this, Display HAL is receiving invalid handles resulting in
framework reboot.

Fix is to cleanup of displaydata session as part of Displaydevice
teardown.

Change-Id: I9d6b5b83a0ccbd92be4998a1a85600b11d594b73