SF CaptureScreen: Add eglMakeCurrent NULL before eglDestroySurface
Refine commit 23dacde4b624921df4f2c9646d20ca7acae357b6

When eglDestroySurface, the surface reference count minus one, but the
surface is not actually destroyed. Only when the next composition happens
and eglMakeCurrent is called to bind to another surface, this old surface
would actually destroy. But now the IGraphicBufferProducer wrapper is exit
already.

The solution is to add make current NULL before destroy surface, to make
sure the surface is truely be destroyed.

Change-Id: I83ec564591fea395b23fc60af839624b0ab1b5cb
CRs-fixed: 539313