Revert "SF CaptureScreen: Add eglMakeCurrent NULL before eglDestroySurface"
This reverts commit c4b70c855b3fea159cf702bc49791e604f477a35.

EGL_BAD_MATCH is generated if draw or read are not compatible
with context. That is, if draw or read are set to EGL_NO_SURFACE,
context must be set to EGL_NO_CONTEXT.

But if context is set to EGL_NO_CONTEXT, checking EGL state would
fail and launcher crash happened.

CRs-fixed: 554939