SurfaceFlinger: Use computed DPI value based on panel dimensions
The DPI value is computed based on the panel's resolution and
physical dimensions. If the panel's physical dimensions are
available, use the computed value. Otherwise, default back to
the one specified in the build property. Also, ensure that the
computed DPI value is rounded off to the nearest valid DPI value
listed in the Android CTS.

Change-Id: I096ecb4a5ad2ea1b21a5bbb6efca8534b871dcd3