SurfaceFlinger: disconnect surface after captureScreenImplLocked
In current sw, We create a IGraphicBufferProducer wrapper that
forwards all calls to the calling binder thread, where they are
executed. This allows the calling thread to be reused (on the other
side) and not depend on having "enough" binder threads to handle the
requests.

But this is not applicable to BufferQueue::disconnect, since it is
executed by eglMakeCurrent, which is only called in next composition
in SurfaceFlinger working thread. This change moves makeCurrent ahead
to delegate disconnect to a binder thread as well.

Change-Id: I715e18329b6574e6422cb3580539679fab474bac
CRs-fixed: 663501